+++ src/ModularPipelines.Build/ReleaseNotes.md
@@ -1,1 +1,2 @@
 *   If no collapsible log mechanisms detected, still segregate logs by a simple text separator
+*   `[SkipIfNoGitHubToken]` attribute, useful on modules that will fail without access to the GitHub token, which can happen for PRs from non-maintiners or dependabot